moving error decorators up to the component_interface class
